Revert "Move pb2 Files to Own Folder"
